refactor(mentor): AllProductArea / BestProductArea 폴더 구조 변경
