+++ src/pages/ItemsPage/components/BestProductArea/BestProductArea.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as var;
+@use "../../../../styles/variables" as var;
 
 .bestProductArea {
   &__title {
